feat (test): a first set of (integration) tests
Tests in `spec/test.js` are written for use with the tape package.
Some files for testing uploads and other functions are stored in
`spec/files`.

ISSUES: A running eXistDB is required in order to have the tests run.
It may be necessary to adjust the connection settings to match your
setup.