update(deps): update node-exist and remove mime
BREAKING CHANGE: `getMimeTypes()` was replaced by `getMimeType(path)`
This was necessary, because the property `mime.types`, is now private.
Also, there is not really a need to get
the complete list of all defined mime-types.

gulp-exist now depends on the latest, scoped version of node-exist.

Since node-exist already exposes all the necessary functionality of mime
and having to instances of it is also error-prone, it was removed.